pyproject.toml
[project]
name = "lucy"
version = "0.1.0"
description = "LUCY voice server"
requires-python = ">=3.11"
dependencies = [
  "livekit-agents[mistralai,openai,silero,turn-detector]~=1.5",
  "python-dotenv",
  "tavily-python",
  "fastapi",
  "uvicorn",
  "livekit-api",
]
